all in usize

+++ rust/src/chat/src/event.rs
@@ -203,7 +203,7 @@ pub enum ChatEvent {
         /// Number of output tokens generated.
         output_token_count: usize,
         /// Number of prompt tokens served from cache.
-        cached_token_count: u32,
+        cached_token_count: usize,
         finish_reason: FinishReason,
         /// Connector-specific KV transfer parameters for disaggregated serving.
         kv_transfer_params: Option<serde_json::Value>,
