+++ rust/src/text/src/output/mod.rs
@@ -24,7 +24,7 @@ pub struct CollectedTextOutput {
     pub token_ids: Vec<u32>,
     pub finish_reason: FinishReason,
     /// Number of prompt tokens served from cache.
-    pub cached_token_count: u32,
+    pub cached_token_count: usize,
     /// Connector-specific KV transfer parameters for disaggregated serving.
     pub kv_transfer_params: Option<serde_json::Value>,
 }
